+++ bfabric/src/bfabric/_oauth/webapp_client.py
@@ -5,8 +5,6 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from bfabric._oauth._constants import DEFAULT_OAUTH_SCOPE
-
 if TYPE_CHECKING:
     from pathlib import Path
 
@@ -35,7 +33,7 @@ class WebappClient:
         *,
         client_id: str,
         client_secret: str,
-        scope: str = DEFAULT_OAUTH_SCOPE,
+        scope: str,
         user_token_cache_path: Path | None = None,
         service_token_cache_path: Path | None = None,
     ) -> WebappClient:
@@ -80,6 +78,7 @@ class WebappClient:
             client_id=client_id,
             client_secret=client_secret,
             token_url=token_url,
+            scope="",
             token=token_dict,
             grant_type="refresh_token",
             token_cache_path=user_token_cache_path,